% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeveloperAppKeys.R
\name{getDeveloperAppKeys}
\alias{getDeveloperAppKeys}
\title{Gets all developer application keys from Betfair.}
\usage{
getDeveloperAppKeys(suppress = FALSE, sslVerify = TRUE)
}
\arguments{
\item{suppress}{Boolean. By default, this parameter is set to FALSE, meaning
that a warning is posted when the getDeveloperAppKeys call throws an error. Changing
this parameter to TRUE will suppress this warning.}

\item{sslVerify}{Boolean. This argument defaults to TRUE and is optional. In
some cases, where users have a self signed SSL Certificate, for example
they may be behind a proxy server, Betfair will fail login with "SSL
certificate problem: self signed certificate in certificate chain". If this
error occurs you may set sslVerify to FALSE. This does open a small
security risk of a man-in-the-middle intercepting your login credentials.}
}
\value{
Response from Betfair is stored in devKeys variable which is a list and is
  returned in that form.
}
\description{
\code{getDeveloperAppKeys} gets all application keys owned by the given developer/vendor and
returns them as a list.
}
\section{Note on \code{devKeysOps} variable}{
 The \code{devKeysOps} variable
  is used to firstly build an R data frame containing all the data to be
  passed to Betfair, in order for the function to execute successfully. The
  data frame is then converted to JSON and included in the HTTP POST request.
  If the getDeveloperAppKeys call throws an error, a data frame containing error
  information is returned.
}

\examples{
\dontrun{
getDeveloperAppKeys() # without any arguments will return details as a data frame
# e.g. getDeveloperAppKeys()[[1]]$appId gives the appId for the first developer key.

}

}
\seealso{
\url{https://docs.developer.betfair.com/display/1smk3cen4v3lu3yomq5qye0ni/getDeveloperAppKeys}
for general information on calling getDeveloperAppKeys via the Betfair API.

\code{\link{loginBF}}, which must be executed first, as this
  function requires a valid session token.
}
