% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listEvents.R
\name{listEvents}
\alias{listEvents}
\title{Return listEvents data}
\usage{
listEvents(eventTypeIds, competitionIds = NULL,
  fromDate = (format(Sys.time(), "\%Y-\%m-\%dT\%TZ")),
  toDate = (format(Sys.time() + 86400, "\%Y-\%m-\%dT\%TZ")),
  sslVerify = TRUE)
}
\arguments{
\item{eventTypeIds}{eventTypeIds String. Restrict markets by event type
associated with the market. (i.e., Football, Horse Racing, etc). Accepts
multiple IDs (See examples). IDs can be obtained via
\code{\link{listEventTypes}}, Required. No default.}

\item{competitionIds}{competitionIds String. Restrict markets by competition
associated with the market (e.g. "81" for Italian football Serie A).
Accepts multiple IDs.}

\item{fromDate}{The start date from which to return matching events. Format
is \%Y-\%m-\%dT\%TZ. Optional. If not defined defaults to current system
date and time.}

\item{toDate}{The end date to stop returning matching events. Format is
\%Y-\%m-\%dT\%TZ. Optional. If not defined defaults to current system
date and time plus 24 hours.}

\item{sslVerify}{Boolean. This argument defaults to TRUE and is optional. In
some cases, where users have a self signed SSL Certificate, for example
they may be behind a proxy server, Betfair will fail login with "SSL
certificate problem: self signed certificate in certificate chain". If this
error occurs you may set sslVerify to FALSE. This does open a small
security risk of a man-in-the-middle intercepting your login credentials.}
}
\value{
Response from Betfair is stored in listEvents variable, which is then
  parsed from JSON as a list. Only the first item of this list contains the
  required event type identification details.
}
\description{
\code{listEvents} returns data corresponding to a specific event type.
}
\details{
\code{listEvents} returns data corresponding to a specific event type, e.g.
Horse Racing, between two date and time stamps. Useful for finding specific
event identification numbers, which are then usually passed to further
functions. Additional returned data includes event name, event country code,
event timezone, event venue (only Horse Racing markets have venues), market
opening date and market count.
}
\section{Note on \code{listEventsOps} variable}{
 The \code{listEventsOps}
  variable is used to firstly build an R data frame containing all the data
  to be passed to Betfair, in order for the function to execute successfully.
  The data frame is then converted to JSON and included in the HTTP POST
  request.
}
\examples{
\dontrun{
# Return all Horse Racing events, between the current date & time and 11pm on December 10th, 2014
listEvents(eventTypeIds = 7,
           fromDate = format(Sys.time(), "\%Y-\%m-\%dT\%TZ"),
           toDate = "2014-12-15T23:00:00Z"
          )

# Return all Horse Racing events, using the default from and to datestamps.
listEvents(eventTypeIds = 7)

# Return all Horse Racing and Football events, using the default from and to datestamps.
listEvents(eventTypeIds = c("1","7"))

# Return all football events associated to the football competittions "Premier League"
# and "Serie A", using the default from and to datestamps

listEvents(eventTypeIds = "1", competitionIds = c("31","81"))
}

}
\seealso{
\code{\link{loginBF}}, which must be executed first.
}

