% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listEventTypes.R
\name{listEventTypes}
\alias{listEventTypes}
\title{Return listEventTypes data}
\usage{
listEventTypes(sslVerify = TRUE)
}
\arguments{
\item{sslVerify}{Boolean. This argument defaults to TRUE and is optional. In
some cases, where users have a self signed SSL Certificate, for example
they may be behind a proxy server, Betfair will fail login with "SSL
certificate problem: self signed certificate in certificate chain". If this
error occurs you may set sslVerify to FALSE. This does open a small
security risk of a man-in-the-middle intercepting your login credentials.}
}
\value{
Response from Betfair is stored in listEventTypes variable, which
  when parsed from JSON as a list. Only the first item of this list contains
  the required event type identification details.
}
\description{
\code{listEventTypes} simply lists all available event types on the Betfair
exchange.
}
\details{
Useful for finding the relevant event type identification number, which is
then usually passed to further functions.
}
\section{Note on \code{listEventTypeOps} variable}{
 The
  \code{listEventTypeOps} variable is used to firstly build an R data frame
  containing all the data to be passed to Betfair, in order for the function
  to execute successfully. The data frame is then converted to JSON and
  included in the HTTP POST request.
}
\examples{
\dontrun{
listEventTypes()
}

}
\seealso{
\code{\link{loginBF}}, which must be executed first.
}

