% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/listCountries.R
\name{listCountries}
\alias{listCountries}
\title{Return listCountries data}
\usage{
listCountries(eventTypeIds)
}
\arguments{
\item{eventTypeIds}{String. Restrict markets by event type associated with
  the market. (i.e., Football, Horse Racing, etc). Accepts multiple IDs (See
  examples). IDs can be obtained via \code{\link{listEventTypes}}, Required. No
  default.}
}
\value{
Response from Betfair is stored in listCountries variable, which is
  then parsed from JSON as a list. Only the first item of this list contains
  the required country identification codes.
}
\description{
\code{listCountries} returns a list of two letter country codes.
}
\details{
\code{listCountries} returns a list of two letter country codes.
\code{eventTypeIds} is required to limit list to specific event types. e.g.
Horse Racing. Returned data includes country codes and count of available
markets in each country, for the specified \code{eventTypeIds}. A complete
list of valid two letter ISO country codes used by Betfair may be found at
the following location:

\url{http://en.wikipedia.org/wiki/ISO_3166-1_alpha-2#Officially_assigned_code_elements}
}
\section{Note on \code{listCountriesOps} variable}{
 The
  \code{listCountriesOps} variable is used to firstly build an R data frame
  containing all the data to be passed to Betfair, in order for the function
  to execute successfully. The data frame is then converted to JSON and
  included in the HTTP POST request.
}
\examples{
\dontrun{
# Return all current country codes and market depth for Horse Racing.
listCountries(eventTypeIds = 7)

# Return all current country codes and market depth for Football and Horse
Racing
listCountries(eventTypeIds = c("1","7"))
}
}
\seealso{
\code{\link{loginBF}}, which must be executed first.
}

